pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deckbot"
version = "0.1.0"
authors = [
  { name="Ryan Porter", email="ryan@example.com" },
]
description = "DeckBot: A Vibe-Coded Presentation Assistant"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "rich",
    "behave",
    "google-generativeai",
    "python-dotenv",
    "requests",
    "Pillow",
    "flask",
    "pyyaml"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/deckbot"

[project.scripts]
deckbot = "vibe_presentation.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
